enhance: Seperate dokan rest controller for role-wise API implementation. (#2484)
* enhance: Split dokan rest controller class for role-wise API implementation.

* fix: Update rest base for admin, vendor, customer controllers.

* fix: api documentation for admin controller usecase.

* fix: update base api class name `DokanBaseRESTController` to `DokanRESTBaseController`, remove redunded properties, update api docs.
