+++ includes/Abstracts/DokanRESTAdminController.php
@@ -2,8 +2,6 @@
 
 namespace WeDevs\Dokan\Abstracts;
 
-use WP_REST_Controller;
-
 /**
 * Admin Dashboard
 *
@@ -11,7 +9,7 @@ use WP_REST_Controller;
 *
 * @package dokan
 */
-abstract class DokanRESTAdminController extends WP_REST_Controller {
+abstract class DokanRESTAdminController extends DokanRESTBaseController {
 
     /**
      * Endpoint namespace.
@@ -21,58 +19,13 @@ abstract class DokanRESTAdminController extends WP_REST_Controller {
     protected $namespace = 'dokan/v1/admin';
 
     /**
-     * Perform permission checking
+     * Check if user has admin permission.
      *
      * @since 2.8.0
      *
-     * @return void
+     * @return bool
      */
     public function check_permission() {
         return current_user_can( 'manage_woocommerce' );
-    }
-
-    /**
-     * Format item's collection for response
-     *
-     * @param  object $response
-     * @param  object $request
-     * @param  array $items
-     * @param  int $total_items
-     *
-     *  @since 2.9.8
-     *
-     * @return object
-     */
-    public function format_collection_response( $response, $request, $total_items ) {
-        if ( $total_items === 0 ) {
-            return $response;
-        }
-
-        // pagation values for headers
-        $per_page = (int) ( ! empty( $request['per_page'] ) ? $request['per_page'] : 20 );
-        $page     = (int) ( ! empty( $request['page'] ) ? $request['page'] : 1 );
-
-        $response->header( 'X-WP-Total', (int) $total_items );
-
-        $max_pages = ceil( $total_items / $per_page );
-
-        $response->header( 'X-WP-TotalPages', (int) $max_pages );
-        $base = add_query_arg( $request->get_query_params(), rest_url( sprintf( '/%s/%s', $this->namespace, $this->base ) ) );
-
-        if ( $page > 1 ) {
-            $prev_page = $page - 1;
-            if ( $prev_page > $max_pages ) {
-                $prev_page = $max_pages;
-            }
-            $prev_link = add_query_arg( 'page', $prev_page, $base );
-            $response->link_header( 'prev', $prev_link );
-        }
-        if ( $max_pages > $page ) {
-            $next_page = $page + 1;
-            $next_link = add_query_arg( 'page', $next_page, $base );
-            $response->link_header( 'next', $next_link );
-        }
-
-        return $response;
     }
 }
